LayoutTests/http/tests/xmlhttprequest/post-blob-content-type-sync-expected.txt

CONSOLE MESSAGE: line 44: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: line 44: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-allow-lists.php due to access control checks.
Test verifies that content MIME type is set correctly when Blob is sent using XMLHttpRequest synchronously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS expectedMimeType is "text/plain;charset=utf-8"
PASS expectedMimeType is ""
PASS expectedMimeType is ""
PASS expectedMimeType is ""
PASS expectedMimeType is ""
PASS expectedMimeType is ""
PASS expectedMimeType is ""
PASS expectedMimeType is "multipart/mixed;boundary=\"--blob-boundary"
PASS expectedMimeType is "multipart/mixed;boundary=\"--blob-boundary\""
PASS expectedMimeType is ""
PASS expectedMimeType is ""
PASS Exception should be thrown.
PASS expectedMimeType is "text/plain;charset=utf-8"
PASS successfullyParsed is true

TEST COMPLETE
